Moved data service tests to a single fixture for baseDataServiceBehavior. Since this code is shared now, it makes sense to have shared tests. The base and base singleton data sources now essentially just wrap the behavior with syntactic sugar.
Tests fail, highlighting an issue with the update that Danny found. It should return the data, not the request object.
